requirements.txt
timm
einops
pthflops
torchsummary
thop
